feat: add type definitions for collections, documents, and query options in JasonDB
