+++ src/types/utils.ts
@@ -1,0 +1,6 @@
+/**
+ * Represents a validation function type.
+ *
+ * @template T The type of data being validated.
+ */
+export type ValidationFunction<T> = (item: T) => boolean;
